+++ CHANGELOG.md
@@ -2,14 +2,26 @@
 
 ## [Unreleased]
 
+### Added
+
+- IP address and ASN are now click-to-copy in the status display
+
+### Fixed
+
 - State updates to `PUT /state` now save correctly (was writing Promise object
   instead of actual data) in [#42](https://github.com/t-mart/mousehole/pull/42)
 - State directory is now created automatically on first write if it doesn't
   exist in [#44](https://github.com/t-mart/mousehole/pull/44)
 - Non-Error exceptions are now properly wrapped in custom error types with
   context in [#45](https://github.com/t-mart/mousehole/pull/45)
+
+### Changed
+
 - Refactor MAM cookie parsing to use `getSetCookie()` in
   [#46](https://github.com/t-mart/mousehole/pull/46)
+
+### Meta
+
 - Start this changelog in [#43](https://github.com/t-mart/mousehole/pull/43)
 
 ## [0.2.0](https://github.com/t-mart/mousehole/releases/tag/v0.2.0) - 2025-09-09